fix(nav): give the merge and duplicates pages their account menu back
Both pages rendered <Navigation currentPath="..." /> and nothing else. The
user menu, the drawer's identity block and the drawer's Settings,
Documentation and Sign out rows are all gated on userEmail reaching
Navigation, so on these two pages there was no way to reach settings or sign
out at any width.

Both were client components, which is why they had no session to pass. Each is
now split the way every other page in the app is built: a server page that
awaits auth() and renders the nav, with the interactive body moved into a
client component beside it.

That also closes a second gap. Nothing enforces auth in the proxy, which only
logs, so every other page gates itself with an auth() check and a redirect to
/login. These two never did; they rendered their chrome to anyone and relied
on the API returning 401 for the data.

The new test walks app/ for pages rendering the nav and asserts each one
passes the account props, so the next page to forget them fails in CI rather
than shipping a drawer with no way out.
